Factor out error crate
