+++ rust/Cargo.toml
@@ -1,9 +1,11 @@
 [workspace]
 resolver = "2"
 members = [
+    "kcl-api",
     "kcl-bumper",
     "kcl-derive-docs",
     "kcl-directory-test-macro",
+    "kcl-error",
     "kcl-language-server",
     "kcl-language-server-release",
     "kcl-lib",
