+++ rust/kcl-error/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "kcl-error"
+version = "0.1.0"
+edition = "2024"
+description = "KCL error definitions"
+
+[dependencies]
+miette = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+schemars = { version = "0.8.17", features = [
+  "impl_json_schema",
+  "indexmap2",
+  "url",
+  "uuid1",
+  "preserve_order",
+] }
+tower-lsp = { workspace = true, features = ["proposed", "default"] }
+ts-rs = { version = "11.0.1", features = [
+  "uuid-impl",
+  "url-impl",
+  "chrono-impl",
+  "indexmap-impl",
+  "no-serde-warnings",
+  "serde-json-impl",
+] }
+
+[lints]
+workspace = true
