+++ rust/kcl-lib/Cargo.toml
@@ -66,6 +66,7 @@ image = { version = "0.25.6", default-features = false, features = ["png"] }
 indexmap = { workspace = true, features = ["serde", "rayon"] }
 itertools = "0.14.0"
 kcl-derive-docs = { version = "0.1", path = "../kcl-derive-docs" }
+kcl-error = { version = "0.1", path = "../kcl-error" }
 kittycad = { workspace = true }
 kittycad-modeling-cmds = { workspace = true }
 lazy_static = { workspace = true }
