+++ rust/kcl-lib/src/modules.rs
@@ -1,46 +1,17 @@
 use std::fmt;
 
 use anyhow::Result;
-use schemars::JsonSchema;
+pub use kcl_error::ModuleId;
 use serde::{Deserialize, Serialize};
 
 use crate::{
+    SourceRange,
     errors::{KclError, KclErrorDetails},
     exec::KclValue,
     execution::{EnvironmentRef, ModuleArtifactState, PreImportedGeometry, typed_path::TypedPath},
     fs::{FileManager, FileSystem},
     parsing::ast::types::{ImportPath, Node, Program},
-    source_range::SourceRange,
 };
-
-/// Identifier of a source file.  Uses a u32 to keep the size small.
-#[derive(
-    Debug, Default, Ord, PartialOrd, Eq, PartialEq, Clone, Copy, Hash, Deserialize, Serialize, ts_rs::TS, JsonSchema,
-)]
-#[ts(export)]
-pub struct ModuleId(u32);
-
-impl ModuleId {
-    pub fn from_usize(id: usize) -> Self {
-        Self(u32::try_from(id).expect("module ID should fit in a u32"))
-    }
-
-    pub fn as_usize(&self) -> usize {
-        usize::try_from(self.0).expect("module ID should fit in a usize")
-    }
-
-    /// Top-level file is the one being executed.
-    /// Represented by module ID of 0, i.e. the default value.
-    pub fn is_top_level(&self) -> bool {
-        *self == Self::default()
-    }
-}
-
-impl std::fmt::Display for ModuleId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
 
 #[derive(Debug, Clone, Default)]
 pub(crate) struct ModuleLoader {
